fix: problems and names of db

+++ backend/src/modules/student/model.py
@@ -1,5 +1,5 @@
 from src.core.database import Base
-from sqlalchemy import String, Numeric
+from sqlalchemy import String, Integer
 from sqlalchemy.orm import Mapped, mapped_column
 
 
@@ -7,4 +7,4 @@ class Student(Base):
     __tablename__ = "students"
     id: Mapped[int] = mapped_column(primary_key=True, index=True)
     name: Mapped[str] = mapped_column(String, nullable=False)
-    age : Mapped[int] = mapped_column(Numeric, nullable=False)    
+    age : Mapped[int] = mapped_column(Integer, nullable=False)
